backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
alembic==1.13.1
pytest==8.0.0
httpx==0.26.0
python-dateutil==2.8.2
cachetools==5.3.2
